pyproject.toml
[project]
name = "probely"
version = "0.0.1"
description = "CLI and SDK wrappers for Probely's API"
authors = [
    { name = "João Silva", email = "joao.silva@probely.com" },
    { name = "Vladimir Kontic", email = "vladimir@probely.com" },
    { name = "Temidire Adams", email = "temidire@probely.com" }
]
license = { file = "LICENSE" }
keywords = ["probely", "cli", "sdk", "api", "client", "wrapper"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
]

readme = "README.md"
requires-python = ">= 3.8"

dependencies = [
    "requests",
    "rich",
    "rich-argparse",
    "environs",
    "marshmallow",
    "pyyaml",
    "mergedeep",
    "python-dateutil", # Python <= 3.10's fromisoformat is not a full ISO-8601 format parser
    "toml", # Python < 3.11's doesn't support toml
]

[project.optional-dependencies]
test = [
    "pytest",
    #"pytest-cov",
    "tox",
]
docs = [
    "sphinx",
    "sphinx-argparse-cli",
    "sphinx-markdown-builder",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
include_trailing_comma = true
line_length = 88

[project.scripts]
probely = "probely.cli:app"

[project.urls]
Homepage = "https://probely.com"
Documentation = "https://developers.probely.com"
Repository = "https://github.com/probely/probely_cli.git"
"Bug Tracker" = "https://github.com/probely/probely_cli/issues"
#Changelog = "https://github.com/probely/probely_cli/blob/master/CHANGELOG.md"




[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

#[tool.hatch.build.targets.wheel]
#packages = ["src/probely"]
